requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
numpy>=1.21.0
opencv-python>=4.5.3.56
Pillow>=8.3.1

# ML dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
huggingface-hub[inference]>=0.19.0

# Utilities
requests>=2.26.0
tqdm>=4.62.2